Renaming folder.
